draft: move ts and rust to their own dir
