+++ packages/coln-js-runtime/Cargo.toml
@@ -17,4 +17,5 @@ tsify = "0.5.6"
 wasm-bindgen = "0.2.123"
 
 [lib]
+path = "src/rust/lib.rs"
 crate-type = ["cdylib", "rlib"]
