+++ packages/coln-js-runtime/src/rust/error.rs
